requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.18
google-generativeai==0.8.3
pymssql==2.3.1
openpyxl==3.1.5
python-dotenv==1.0.1
